requirements.txt
streamlit
requests
beautifulsoup4
lxml
langcodes
language-data
rapidfuzz
asyncwhois
openai
Unidecode
gspread
google-auth
google-auth-oauthlib


